% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_box_plots.R
\name{check_numeric_columns}
\alias{check_numeric_columns}
\title{Returns a data frame containing only the numeric columns of the original data frame}
\usage{
check_numeric_columns(data)
}
\arguments{
\item{data}{Any data frame.}
}
\value{
A subset of numeric variables from the original data frame.
}
\description{
Returns a data frame containing only the numeric columns of the original data frame
}
