% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar_plot.R
\name{create_bar_plots}
\alias{create_bar_plots}
\title{Creates bar plots of every categorical variable in a data frame}
\usage{
create_bar_plots(data)
}
\arguments{
\item{dataframe}{Any data frame}
}
\value{
A ggplot made bar plot
}
\description{
Creates bar plots of every categorical variable in a data frame
}
