% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_box_plots.R
\name{create_box_plots}
\alias{create_box_plots}
\title{Creates box plots of every numeric variable in any data frame.}
\usage{
create_box_plots(data)
}
\arguments{
\item{dataframe}{Any data frame}
}
\value{
A ggplot graph of box plots.
}
\description{
Creates box plots of every numeric variable in any data frame.
}
