% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_plot.R
\name{create_line_plot}
\alias{create_line_plot}
\title{Creates a line plot of two quantitative variables}
\usage{
create_line_plot(data, x, y)
}
\arguments{
\item{x}{First selected quantitative variable on x axis}

\item{y}{Second selected quantitative variable on y axis}

\item{dataframe}{Any data frame}
}
\value{
A ggplot graph of a line plot.
}
\description{
Creates a line plot of two quantitative variables
}
