% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Creates a scatter plot of two different quantitative variables}
\usage{
create_scatter_plot(data, x, y)
}
\arguments{
\item{x}{First selected quantitative variable on x axis}

\item{y}{Second selected quantitative variable on y axis}

\item{dataframe}{Any data frame}
}
\value{
A ggplot graph of a scatter plot.
}
\description{
Creates a scatter plot of two different quantitative variables
}
