% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar_plot.R
\name{filter_categorical}
\alias{filter_categorical}
\title{Returns a data frame containing only the categorical columns of the original data frame}
\usage{
filter_categorical(data)
}
\arguments{
\item{data}{Any data frame.}
}
\value{
A subset of numeric variables from the original data frame.
}
\description{
Returns a data frame containing only the categorical columns of the original data frame
}
