% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_plot.R, R/create_scatter_plot.R
\name{numeric_columns}
\alias{numeric_columns}
\title{Returns a data frame containing numeric columns of the original data frame}
\usage{
numeric_columns(data)

numeric_columns(data)
}
\arguments{
\item{data}{Any data frame.}
}
\value{
A subset of numeric variables from the original data fram to test if line plot can be produced

A subset of numeric variables from the original data frame. Used to test to see if scatter plot can be made
}
\description{
Returns a data frame containing numeric columns of the original data frame

Returns a data frame containing numeric columns of the original data frame
}
